+++ tests/testthat/test-estimate_dispersion.R
@@ -79,6 +79,21 @@ test_that("Error conditions are properly handled", {
   expect_error(estimate_dispersion(valid_nowcasts, valid_trunc_rts, n = -1))
   expect_error(estimate_dispersion(valid_nowcasts, valid_trunc_rts, n = 1.5))
   expect_error(estimate_dispersion(valid_nowcasts, valid_trunc_rts, n = 3))
+
+  # pt nowcast contains NAs (use trunc rts) or is empty
+  expect_error(estimate_dispersion(valid_trunc_rts, valid_trunc_rts))
+  expect_error(estimate_dispersion(list(), valid_trunc_rts))
+
+  # trunc rep mat list does not contain NAs (use pt nowcasts)
+  expect_error(estimate_dispersion(valid_nowcasts, list()))
+
+  # observations contain non-integers
+  test_triangle_decimal <- test_triangle + 0.1
+  non_integer_trunc_rts <- list(
+    test_triangle_decimal[1:5, ],
+    test_triangle_decimal[1:4, ]
+  )
+  expect_error(estimate_dispersion(valid_nowcasts, non_integer_trunc_rts))
 })
 
 ### Test 4: Edge Cases ---------------------------------------------------------
@@ -100,9 +115,8 @@ test_that("Matrix dimension validation works", {
     test_triangle[1:3, ]
   )
   expect_error(
-    estimate_dispersion(valid_nowcasts, bad_trunc_rts),
-    "Dimensions of the first `n` matrices in `pt_nowcast_mat_list` and"
-  ) # nolint
+    estimate_dispersion(valid_nowcasts, bad_trunc_rts)
+  )
 })
 
 ## Test 6: fit_nb returns NA if nothing passed to it---------------------------